pyproject.toml
[project]
name = "mwareeth"
version = "0.0.0"
description = "KBS system to divide estate according to Islamic law."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "babel>=2.17.0",
    "graphviz>=0.21",
    "python-statemachine>=2.5.0",
    "tk>=0.1.0",
]

[project.optional-dependencies]
gui = [
    "pillow>=10.0.0",  # For advanced image handling in the GUI
]

[project.scripts]
mwareeth = "main:main"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "pyre-check>=0.9.25",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
    "setuptools>=80.9.0",
]
